requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.15
langchain-groq>=0.1.0
langchain-huggingface>=0.0.2
chromadb>=0.4.22
python-dotenv>=1.0.0
streamlit>=1.27.0
tavily-python>=0.3.0
pypdf>=3.17.0
langsmith>=0.0.72
bs4>=0.0.1
requests>=2.31.0
sentence-transformers==2.5.1
torch==2.2.1